7-16 add game.h and game.cpp move functionality to game from engine

+++ Engine/renderer.cpp
@@ -43,7 +43,7 @@ void Renderer::DrawTexture(SDL_Texture * texture, const Vector2D & position, con
 	SDL_Rect dest = { point.x, point.y, 0, 0 };
 	//SDL_Rect dest = { (point.x * scale.x), (point.y * scale.y), 0, 0 };
 	SDL_QueryTexture(texture, nullptr, nullptr, &dest.w, &dest.h);
-	dest.w *= scale.x;
-	dest.h *= scale.y;
+	dest.w *= static_cast<int>(scale.x);
+	dest.h *= static_cast<int>(scale.y);
 	SDL_RenderCopyEx(m_renderer, texture, nullptr, &dest, angle, nullptr, SDL_FLIP_NONE);
 }
